refactor(recovery): move recovery package to feature/recovery
Atomic relocation of the recovery boundary out of app/app:

- 8 source files moved to feature/recovery (domain/, diagnostics/, root)
- 2 unit tests moved with their packages updated
- 11 strings (9 recovery_* + 2 renamed recovery_restore_failure_*)
  relocated; English + Russian; intent-side strings renamed from
  app_dialog_restore_failure_* for namespace consistency
- AndroidManifest FQCN updated to feature/recovery package
- BaseApplication + MainActivity imports updated; app/app gains the
  feature/recovery module dependency
- Visibility flips for cross-module access: RestoreRecoveryCoordinator,
  StartupMigrationCoordinator, StartupCheck (sealed iface + variants),
  StartupMigrationFailureReason (referenced by public StartupCheck
  property), and RecoveryActivity flip to public class; nested
  @Inject constructors are kept internal so reporters and observer
  stay internal to feature/recovery
- AppDialogObserverBootstrapEntryPoint relocates to feature/recovery/
  boot/, returning the new RecoveryBootstrap marker interface so the
  observer can stay internal while BaseApplication still drives eager
  construction. The marker + @Binds module preserve the Phase B
  BLOCKER 1 guarantee: subscription registered before any UI dispatch
- DuplicateStrings lint suppressions added for the two share-chooser
  strings that legitimately share text (Scenarios 1 + 2)
- AppDialogHost KDoc reference updated to the new observer location

Verified: assembleDebug + detekt (39 modules incl. recovery) +
:app:dev:lintDebug (0 errors, 28 pre-existing warnings) +
testDebugUnitTest all green.
