+++ feature/app-dialogs/impl/src/main/res/values-ru/strings.xml
@@ -13,8 +13,6 @@
     <string name="app_dialog_restore_failure_confirm" tools:ignore="DuplicateStrings">ОК</string>
     <string name="app_dialog_restore_failure_report_action">Сообщить</string>
     <string name="app_dialog_restore_failure_export_action">Экспорт диагностики</string>
-    <string name="app_dialog_restore_failure_share_chooser">Поделиться диагностикой</string>
-    <string name="app_dialog_restore_failure_report_title">Ошибка восстановления</string>
 
     <!-- UndoRestoreConfirmation -->
     <string name="app_dialog_undo_restore_confirmation_title">Отменить последнее восстановление?</string>
