+++ feature/recovery/src/main/kotlin/io/github/stslex/workeeper/feature/recovery/diagnostics/RecoveryDiagnosticsExporter.kt
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.recovery
+package io.github.stslex.workeeper.feature.recovery.diagnostics
 
 import android.content.Context
 import android.content.pm.PackageManager
