+++ feature/recovery/src/main/kotlin/io/github/stslex/workeeper/feature/recovery/diagnostics/RestoreRecoveryReporter.kt
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.recovery
+package io.github.stslex.workeeper.feature.recovery.diagnostics
 
 import io.github.stslex.workeeper.core.core.logger.FirebaseCrashlyticsHolder
 import io.github.stslex.workeeper.core.data.backup.api.restore.RestoreInProgressContext
