+++ feature/recovery/src/main/kotlin/io/github/stslex/workeeper/feature/recovery/domain/UndoRestoreOutcome.kt
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.recovery
+package io.github.stslex.workeeper.feature.recovery.domain
 
 /**
  * Three-way outcome of [RestoreRecoveryCoordinator.performUndoRestore]. The
