+++ feature/recovery/src/main/res/values-ru/strings.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources xmlns:tools="http://schemas.android.com/tools">
+
+    <string name="recovery_title">Требуется обновление приложения</string>
+    <string name="recovery_body">Файл с вашими данными создан в версии, которую эта сборка не может прочитать. Ваши данные сохранены на устройстве. Обновите приложение или экспортируйте необработанные данные для резервной копии.</string>
+    <string name="recovery_update_app">Обновить приложение</string>
+    <string name="recovery_export_data">Экспорт данных</string>
+    <string name="recovery_report_issue">Сообщить</string>
+    <string name="recovery_export_diagnostics">Экспорт диагностики</string>
+    <string name="recovery_export_data_chooser">Поделиться данными</string>
+    <string name="recovery_export_diagnostics_chooser" tools:ignore="DuplicateStrings">Поделиться диагностикой</string>
+    <string name="recovery_report_title">Ошибка миграции при запуске</string>
+
+    <string name="recovery_restore_failure_share_chooser" tools:ignore="DuplicateStrings">Поделиться диагностикой</string>
+    <string name="recovery_restore_failure_report_title" tools:ignore="DuplicateStrings">Ошибка восстановления</string>
+
+</resources>
