+++ feature/recovery/src/test/kotlin/io/github/stslex/workeeper/feature/recovery/domain/StartupMigrationCoordinatorTest.kt
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.recovery
+package io.github.stslex.workeeper.feature.recovery.domain
 
 import android.content.Context
 import io.github.stslex.workeeper.core.data.backup.api.error.BackupError
@@ -7,6 +7,7 @@ import io.github.stslex.workeeper.core.data.backup.api.result.BackupResult
 import io.github.stslex.workeeper.core.data.database.AppDatabase
 import io.github.stslex.workeeper.core.data.database.migration.APP_DATABASE_VERSION
 import io.github.stslex.workeeper.core.data.database.snapshot.DatabaseSnapshotProvider
+import io.github.stslex.workeeper.feature.recovery.diagnostics.StartupMigrationReporter
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
